+++ CHANGELOG.md
@@ -112,6 +112,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Remove parameter `$options` from `ASTDefinitionBuilder`
 - Remove `FieldDefinition::create()` in favor of `new FieldDefinition()`
 - Remove `GraphQL\Exception\InvalidArgument`
+- Remove `Utils::find()`, `Utils::every()` and `Utils::invariant()`
 
 ## 14.11.5